Refactored grains pass all tests using OLD code path.
