+++ mediagrains/grains/CodedAudioGrain.py
@@ -1,19 +1,20 @@
-from mediatimestamp.immutable import (
-    Timestamp,
-    TimeOffset)
+from uuid import UUID
+from mediatimestamp.immutable import Timestamp, SupportsMediaTimestamp, mediatimestamp, TimeOffset
 from fractions import Fraction
 
 from typing import (
-    Optional)
+    Optional,
+    Sized,
+    cast)
 from ..typing import (
     CodedAudioGrainMetadataDict,
     GrainDataParameterType)
 
 from ..cogenums import CogAudioFormat
-from .Grain import Grain
+from .BaseGrain import BaseGrain
 
 
-class CodedAudioGrain(Grain):
+class CodedAudioGrain(BaseGrain):
     """\
 A class representing a coded audio grain.
 
@@ -90,7 +91,77 @@ priming
 remainder
     An integer
 """
-    def __init__(self, meta: CodedAudioGrainMetadataDict, data: GrainDataParameterType):
+    def __init__(self,
+                 meta: CodedAudioGrainMetadataDict,
+                 data: GrainDataParameterType,
+                 src_id: Optional[UUID] = None,
+                 flow_id: Optional[UUID] = None,
+                 origin_timestamp: Optional[SupportsMediaTimestamp] = None,
+                 creation_timestamp: Optional[SupportsMediaTimestamp] = None,
+                 sync_timestamp: Optional[SupportsMediaTimestamp] = None,
+                 rate: Fraction = Fraction(25, 1),
+                 duration: Fraction = Fraction(1, 25),
+                 cog_audio_format: CogAudioFormat = CogAudioFormat.INVALID,
+                 samples: int = 0,
+                 channels: int = 0,
+                 priming: int = 0,
+                 remainder: int = 0,
+                 sample_rate: int = 48000,
+                 length: Optional[int] = None):
+
+        if length is None:
+            if data is not None and hasattr(data, "__len__"):
+                length = len(cast(Sized, data))
+            else:
+                length = 0
+
+        if meta is None:
+            if src_id is None:
+                raise AttributeError("src_id is None. Meta is None so src_id must not be None.")
+            if flow_id is None:
+                raise AttributeError("flow_id is None. Meta is None so flow_id must not be None.")
+
+            if not isinstance(src_id, UUID):
+                raise AttributeError(f"src_id: Seen type {type(src_id)}, expected UUID.")
+            if not isinstance(flow_id, UUID):
+                raise AttributeError(f"flow_id: Seen type {type(flow_id)}, expected UUID.")
+
+            creation_timestamp = Timestamp.get_time() if not creation_timestamp else creation_timestamp
+            if origin_timestamp is None:
+                origin_timestamp = creation_timestamp
+            if sync_timestamp is None:
+                sync_timestamp = origin_timestamp
+            meta = {
+                "@_ns": "urn:x-ipstudio:ns:0.1",
+                "grain": {
+                    'grain_type': "coded_audio",
+                    'source_id': str(src_id),
+                    'flow_id': str(flow_id),
+                    'origin_timestamp': str(mediatimestamp(origin_timestamp)),
+                    'sync_timestamp': str(mediatimestamp(sync_timestamp)),
+                    'creation_timestamp': str(mediatimestamp(creation_timestamp)),
+                    'rate': {
+                        'numerator': Fraction(rate).numerator,
+                        'denominator': Fraction(rate).denominator
+                    },
+                    'duration': {
+                        'numerator': Fraction(duration).numerator,
+                        'denominator': Fraction(duration).denominator
+                    },
+                    'cog_coded_audio': {
+                        "format": cog_audio_format,
+                        "samples": samples,
+                        "channels": channels,
+                        "priming": priming,
+                        "remainder": remainder,
+                        "sample_rate": sample_rate
+                    }
+                }
+            }
+
+        if data is None:
+            data = bytearray(length)
+
         super().__init__(meta, data)
         self.meta: CodedAudioGrainMetadataDict
